% Auto-generated: do not edit by hand
\name{dashkitTable}

\alias{dashkitTable}

\title{DashkitTable component}

\description{

}

\usage{
dashkitTable(id=NULL, cellClassName=NULL, className=NULL,
colHeaders=NULL, columnSorting=NULL, columns=NULL,
contextMenu=NULL, data=NULL, dropdownMenu=NULL,
filters=NULL, headerClassName=NULL, height=NULL,
licenseKey=NULL, multiColumnSorting=NULL, rowHeaders=NULL,
rowHeight=NULL, settings=NULL, stretchH=NULL,
themeName=NULL, width=NULL)
}

\arguments{
\item{id}{Character. }

\item{cellClassName}{Character. }

\item{className}{Character. }

\item{colHeaders}{Logical | numeric | character | named list | unnamed list. }

\item{columnSorting}{Logical | numeric | character | named list | unnamed list. }

\item{columns}{List of logical | numeric | character | named list | unnamed lists. }

\item{contextMenu}{Logical | numeric | character | named list | unnamed list. }

\item{data}{List of logical | numeric | character | named list | unnamed lists | list of list of logical | numeric | character | named list | unnamed listss. }

\item{dropdownMenu}{Logical | numeric | character | named list | unnamed list. }

\item{filters}{Logical | numeric | character | named list | unnamed list. }

\item{headerClassName}{Character. }

\item{height}{Character | numeric. }

\item{licenseKey}{Character. }

\item{multiColumnSorting}{Logical | numeric | character | named list | unnamed list. }

\item{rowHeaders}{Logical | numeric | character | named list | unnamed list. }

\item{rowHeight}{Numeric. }

\item{settings}{Logical | numeric | character | named list | unnamed list. }

\item{stretchH}{Character. }

\item{themeName}{Character. }

\item{width}{Character | numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

